do apt-key the stupid way
The latest versions of `apt-key`, `gpg`, and/or `gpg2` helpfully and
_strenuously_ insist upon socket communication with `gpg-agent`.

With the introduction of GPG 2, the presence of a running gpg-agent is now
required. Most (if not all) operations normally performed by the `gpg` program
will fail if the client (the program you thought you had) cannot communicate
with the server (the agent GNU foisted upon you). Because of course you wanted
this.

This is a joy if, say, you like to play inside environments that lack a `tty`
and/or permissions to create Unix sockets. The docker container is one such
environment. But, cheer up --- it's only a problem if you want to do something
silly like `apt-key add` a pubkey.

Any troglodyte without a `gpg-agent` should just `gpg --dearmor` their pubkeys
directly into `/etc/apt/trusted.gpg.d/` like the uncultured swine they are.